PRSD-358: move getUserList to LocalAuthorityDataService
